Navbar, language switching button, and navbar animation were placed in separate files.
